Account suggestion based on last entry
Only in cases where hledger-iadd wouldn't suggest any other account, suggest based on the last entered transaction.

Since the transaction isn't necessarily the same type, we can't rely on matching the data so we must use the order. This way if the user typically uses a certain order like expense category and then payment method it works. Useful if entering many similar postings in a row.

My use case for writing this was entering transactions from a credit card statement where the first account was usually food, and the second posting was always my credit card.