Make recent completion work for future dates
Make the recent date completion work correctly for future dates and also
for 'strange past dates' (like 'last november'). Now, an incomplete date
is completed to the most recent day in the past matching that incomplete
specification. If none is found, it completes to the nearest day in the
future matching that definition.

Add more testcases to cover that.