Copy `hledger add` for suggesting accounts/amounts
The previous behavior was to suggest accounts from the last similar
transaction in descending order of their amounts. The suggested
amount for an unbalanced transaction was always the balancing amount.

This had two problems: The accounts weren't in the same order as in
the last transaction and the suggested amount was useless for
transactions with more than two accounts.

Now, the behavior of `hledger add` is followed more closely: Accounts
are always suggested in the same order as in the last similar
transaction. If the accounts of the currently edited transaction are
still a subset of the last transaction, the amounts of this
transaction are used as suggestion. Otherwise, the balancing amount it
used.