Don't save final account answer for undo history
This lets the editor undo history adhere to the current 'undo'
behaviour.

Do not save the empty answer to the last account question, because it
indicates the end of the transaction. Right now, 'undo' jumps straight
back to the last amount question instead of jumpting to the finalizing
account question.

In the future there are two possibilities:

  1) leave the 'undo' behaviour as it is.
  2) let 'undo' jump from the FinalQuestion to the last account
  question.