Implement detection of duplicate transactions
Warns the user if they are trying to insert a transaction that is a
literal duplicate of an already existing one.

The UI is not final at this stage. It displays a "(duplicate)" string
in the "do you want to insert this transaction to the journal" prompt.
In the future the original transaction should be displayed and the
warning should be more prominent (maybe by use of color).